requirements.txt
mutagen
python-dotenv
rapidfuzz
requests
rich
urllib3
